refactor: extract board util module
