+++ packages/board/server.mjs
@@ -38,24 +38,9 @@ import {
   NOTIF_HISTORY_FILE,
   VAPID_SUBJECT,
 } from './lib/config.mjs';
+import { csvCell, originAllowed, eventSurface, readFileSafe } from './lib/util.mjs';
 
 // ── Helpers ────────────────────────────────────────────────────────────────────
-
-/**
-
- */
-/**
- * Escape a single value for inclusion in a CSV cell.
- * Quotes the value if it contains comma / quote / newline.
- */
-function csvCell(v) {
-  if (v == null) return '';
-  const s = typeof v === 'string' ? v : JSON.stringify(v);
-  if (/[",\n]/.test(s)) return '"' + s.replace(/"/g, '""') + '"';
-  return s;
-}
-
-
 
 // ── Project registry ───────────────────────────────────────────────────────────
 function readProjectsRegistry() {
@@ -304,16 +289,6 @@ function broadcast(event, data) {
   }
 }
 
-// Only same-origin (the board's own page) may make a state-changing request — a malicious page must not.
-function originAllowed(req) {
-  const o = req.headers.origin || req.headers.referer || '';
-  if (!o) return true; // same-origin fetch / curl with no Origin
-  // True same-origin: the browser's Origin matches the host this request arrived on
-  // (covers a tunnelled/hosted console at console.client.com, http or https).
-  const self = req.headers.host ? [`http://${req.headers.host}`, `https://${req.headers.host}`] : [];
-  return [`http://localhost:${PORT}`, `http://127.0.0.1:${PORT}`, ...self].some((e) => o === e || o.startsWith(e + '/'));
-}
-
 function broadcastTasks(cwd) {
   const msg = `event: tasks\ndata: ${JSON.stringify(getTasks(cwd))}\n\n`;
   for (const res of sseClients) {
@@ -349,13 +324,6 @@ function saveNotifHistory() {
  * Record a notification, broadcast via SSE, and persist.
  * Called alongside fireEmailAlert / firePushAlert at every trigger point.
  */
-// Which surface an alert belongs to. Operate-side events (autopilot runtime: dead-letters,
-// connector health, case SLA, gate/safe-mode pushes) are the operator console's concern and
-// must not clutter the builder board's notifications — the two surfaces are separated.
-function eventSurface(event) {
-  return /^(autopilot\.|dead-letter|connector\.|sla\.)/.test(String(event || '')) ? 'operate' : 'builder';
-}
-
 function addNotification(event, payload) {
   const notif = {
     id: crypto.randomUUID(),
@@ -379,9 +347,6 @@ function addNotification(event, payload) {
 loadNotifHistory();
 
 // ── Memory: 4-layer file contents ─────────────────────────────────────────────
-function readFileSafe(p) {
-  try { return fs.existsSync(p) ? fs.readFileSync(p, 'utf8') : null; } catch { return null; }
-}
 function getMemory(cwd = process.cwd()) {
   const home = os.homedir();
   const layers = [
